# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/activities": {
      "get": {
        "operationId": "ActivityController_list",
        "summary": "List activities or paginate by lineup artist",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lineupArtistId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BackendActivityDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "activities"
        ]
      }
    },
    "/api/activities/resolve": {
      "get": {
        "operationId": "ActivityController_resolve",
        "summary": "Resolve activity by keyword",
        "parameters": [
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ActivityResolveResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "activities"
        ]
      }
    },
...
    "/api/activities/{legacyId}": {
      "get": {
        "operationId": "ActivityController_getByLegacyId",
        "summary": "Get activity detail by legacyId",
        "parameters": [
          {
            "name": "legacyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BackendActivityDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "activities"
        ]
      },
      "patch": {
        "operationId": "ActivityController_update",
        "summary": "Update activity metadata",
        "parameters": [
          {
            "name": "legacyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BackendActivityDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "activities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/activities/{legacyId}/register": {
      "post": {
        "operationId": "ActivityController_register",
        "summary": "Register current user for activity",
        "parameters": [
          {
            "name": "legacyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ActivityRegistrationResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "activities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ActivityController_unregister",
        "summary": "Unregister current user from activity",
        "parameters": [
          {
            "name": "legacyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ActivityUnregisterResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "activities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/activities/{legacyId}/register/wechat-updates": {
      "post": {
        "operationId": "ActivityController_optInWechatActivityUpdates",
        "summary": "Opt in to WeChat activity update notifications",
        "parameters": [
          {
            "name": "legacyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ActivityWechatUpdateOptInResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "activities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Sync API",
    "description": "Sync App REST API. Responses are wrapped by TransformInterceptor as { code, message, data }.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiEnvelopeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "Response payload"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
      "BackendActivityDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "legacyId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "alias": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "date": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "attendees": {
            "type": "number"
          },
          "travelGuideSupported": {
            "type": "boolean"
          }
        },
        "required": [
          "_id",
          "legacyId",
          "name",
          "code"
        ]
      },
      "ActivityResolveResultDto": {
        "type": "object",
        "properties": {
          "activity": {
            "$ref": "#/components/schemas/BackendActivityDto"
          },
          "keyword": {
            "type": "string"
          }
        }
      },
...
      "UpdateActivityDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "activityType": {
            "type": "string",
            "enum": [
              "festival",
              "indoor"
            ]
          },
          "hot": {
            "type": "boolean"
          },
          "infoSource": {
            "type": "string"
          },
          "infoUpdatedAt": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
      "ActivityRegistrationResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "activityLegacyId": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "example": "registered"
          },
          "alreadyRegistered": {
            "type": "boolean"
          },
          "attendees": {
            "type": "number"
          }
        },
        "required": [
          "ok",
          "activityLegacyId",
          "status",
          "attendees"
        ]
      },
      "ActivityWechatUpdateOptInResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "activityLegacyId": {
            "type": "number"
          },
          "wechatActivityUpdateOptIn": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok",
          "activityLegacyId",
          "wechatActivityUpdateOptIn"
        ]
      },
      "ActivityUnregisterResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "activityLegacyId": {
            "type": "number"
          },
          "wasRegistered": {
            "type": "boolean"
          },
          "attendees": {
            "type": "number"
          }
        },
        "required": [
          "ok",
          "activityLegacyId",
          "attendees"
        ]
      },
...
    }
  }
}